+++ discovery/src/main/java/com/ibm/watson/developer_cloud/discovery/v1/model/UpdateTrainingExampleOptions.java
@@ -24,7 +24,7 @@ public class UpdateTrainingExampleOptions extends GenericModel {
   private String collectionId;
   private String queryId;
   private String exampleId;
-  private int relevance;
+  private Long relevance;
   private String crossReference;
 
   /**
@@ -35,7 +35,7 @@ public class UpdateTrainingExampleOptions extends GenericModel {
     private String collectionId;
     private String queryId;
     private String exampleId;
-    private int relevance;
+    private Long relevance;
     private String crossReference;
 
     private Builder(UpdateTrainingExampleOptions updateTrainingExampleOptions) {
@@ -127,7 +127,7 @@ public class UpdateTrainingExampleOptions extends GenericModel {
      * @param relevance the relevance
      * @return the UpdateTrainingExampleOptions builder
      */
-    public Builder relevance(int relevance) {
+    public Builder relevance(long relevance) {
       this.relevance = relevance;
       return this;
     }
@@ -215,7 +215,7 @@ public class UpdateTrainingExampleOptions extends GenericModel {
    *
    * @return the relevance
    */
-  public int relevance() {
+  public Long relevance() {
     return relevance;
   }
